python/pypi/solana/requirements.txt
solana==0.35.0
python-dotenv
httpx_socks
